Decompose qops to call sites
